1. Бутстрап 2. Типы писем, ящики ЦБ, письма; контроллеры, модельки и вьюхи к ним.

+++ db/migrate/20160621185952_create_type_letters.rb
@@ -3,7 +3,7 @@ class CreateTypeLetters < ActiveRecord::Migration
     create_table :type_letters do |t|
       # 1 - обычное письмо, 2 - приказ, 9 - перечень приказов, 51 - курсы, 52  - ответ на письмо ЦБ, 53 - письмо в ЦБ
       t.string  :direction  # 1 - вх, 2 - исх
-      t.float   :service_number # пример: 1
+      t.integer :service_number # пример: 1
       t.string  :name           # пример: обычное письмо
 
       t.timestamps null: false
